requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cpu
--pre
torch
torchvision
torchaudio
uvicorn[standard]
fastapi
pillow
transformers
python-multipart
sentencepiece